Make meal plan

+++ DietScoopApp/app/src/main/java/com/example/dietscoop/Adapters/MealDayRecyclerAdapter.java
@@ -12,6 +12,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.example.dietscoop.Activities.RecyclerItemClickListener;
 import com.example.dietscoop.Data.FoodItem;
 import com.example.dietscoop.Data.Ingredient.Ingredient;
+import com.example.dietscoop.Data.Ingredient.IngredientInMealDay;
 import com.example.dietscoop.Data.Recipe.RecipeInMealDay;
 import com.example.dietscoop.R;
 
@@ -29,6 +30,7 @@ public class MealDayRecyclerAdapter extends RecyclerView.Adapter<MealDayRecycler
 
         private final TextView mealDescription;
         private final TextView mealQuantity;
+//        private final TextView mealUnit;
 
         public MealHolder(View view) {
             super(view);
@@ -36,6 +38,7 @@ public class MealDayRecyclerAdapter extends RecyclerView.Adapter<MealDayRecycler
             //View retrieval for each individual item in Recycler View:
             mealDescription = (TextView) view.findViewById(R.id.meal_in_mealday_description);
             mealQuantity = (TextView) view.findViewById(R.id.meal_in_mealday_quantity_or_servings);
+//            mealUnit = (TextView) view.findViewById(R.id.meal_in_day_unit);
             view.setOnClickListener(this); //Setting up the onClick Listener:
         }
 
@@ -47,6 +50,8 @@ public class MealDayRecyclerAdapter extends RecyclerView.Adapter<MealDayRecycler
         public TextView getMealQuantity() {
             return mealQuantity;
         }
+
+//        public TextView getMealUnit() { return mealUnit; }
 
         @Override
         public void onClick(View view) {
@@ -88,7 +93,7 @@ public class MealDayRecyclerAdapter extends RecyclerView.Adapter<MealDayRecycler
             String quantityUnit = ((Ingredient)meals.get(position)).getMeasurementUnit().toString();
             holder.getMealQuantity().setText(String.valueOf(((Ingredient)meals.get(position)).getAmount()) + " " + quantityUnit);
         } else if (meals.get(position).getType().equals("Recipe")) {
-            Integer numOfServings = ((RecipeInMealDay)meals.get(position)).getDesiredNumOfServings();
+            Double numOfServings = ((RecipeInMealDay)meals.get(position)).getDesiredNumOfServings();
             holder.getMealQuantity().setText(String.valueOf(numOfServings) + " Servings");
         }
